refator(darwin): share code between iOS & macOS
