+++ media_kit_video/common/darwin/Classes/ResizableTextureProtocol.swift
