+++ media_kit_video/common/darwin/Classes/SwappableObjectManager.swift
