--- media_kit_video/ios/Classes/MediaKitVideoPlugin.swift
@@ -1,113 +1,0 @@
-#if canImport(UIKit)
-  import UIKit
-#elseif canImport(Cocoa)
-  import Cocoa
-#endif
-
-#if canImport(Flutter)
-  import Flutter
-#elseif canImport(FlutterMacOS)
-  import FlutterMacOS
-#endif
-
-public class MediaKitVideoPlugin: NSObject, FlutterPlugin {
-  private static let CHANNEL_NAME = "com.alexmercerind/media_kit_video"
-
-  public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(
-      name: CHANNEL_NAME,
-      binaryMessenger: registrar.messenger()
-    )
-    let instance = MediaKitVideoPlugin.init(
-      registry: registrar.textures(),
-      channel: channel
-    )
-    registrar.addMethodCallDelegate(instance, channel: channel)
-  }
-
-  private let channel: FlutterMethodChannel
-  private let videoOutputManager: VideoOutputManager
-
-  init(registry: FlutterTextureRegistry, channel: FlutterMethodChannel) {
-    self.channel = channel
-    self.videoOutputManager = VideoOutputManager(
-      registry: registry
-    )
-  }
-
-  public func handle(
-    _ call: FlutterMethodCall,
-    result: @escaping FlutterResult
-  ) {
-    switch call.method {
-    case "VideoOutputManager.Create":
-      handleCreateMethodCall(call.arguments, result)
-      break
-    case "VideoOutputManager.Dispose":
-      handleDisposeMethodCall(call.arguments, result)
-      break
-    default:
-      result(FlutterMethodNotImplemented)
-      break
-    }
-  }
-
-  private func handleCreateMethodCall(
-    _ arguments: Any?,
-    _ result: FlutterResult
-  ) {
-    let args = arguments as? [String: Any]
-    let handleStr = args?["handle"] as! String
-    let widthStr = args?["width"] as! String
-    let heightStr = args?["height"] as! String
-    let enableHardwareAcceleration =
-      args?["enableHardwareAcceleration"] as! Bool
-
-    let handle: Int64? = Int64(handleStr)
-    let width: Int64? = Int64(widthStr)
-    let height: Int64? = Int64(heightStr)
-
-    assert(handle != nil, "handle must be an Int64")
-
-    self.videoOutputManager.create(
-      handle: handle!,
-      width: width,
-      height: height,
-      enableHardwareAcceleration: enableHardwareAcceleration,
-      textureUpdateCallback: { (_ textureId: Int64, _ size: CGSize) -> Void in
-        self.channel.invokeMethod(
-          "VideoOutput.Resize",
-          arguments: [
-            "handle": handle!,
-            "id": textureId,
-            "rect": [
-              "top": 0,
-              "left": 0,
-              "width": size.width,
-              "height": size.height,
-            ],
-          ]
-        )
-      }
-    )
-
-    result(nil)
-  }
-
-  private func handleDisposeMethodCall(
-    _ arguments: Any?,
-    _ result: FlutterResult
-  ) {
-    let args = arguments as? [String: Any]
-    let handleStr = args?["handle"] as! String
-    let handle: Int64? = Int64(handleStr)
-
-    assert(handle != nil, "handle must be an Int64")
-
-    self.videoOutputManager.destroy(
-      handle: handle!
-    )
-
-    result(nil)
-  }
-}
