+++ media_kit_video/ios/Classes/common/SwappableObjectManager.swift
@@ -1,0 +1,1 @@
+../../../common/darwin/Classes/SwappableObjectManager.swift
