+++ media_kit_video/ios/Classes/common/VideoOutputManager.swift
@@ -1,0 +1,1 @@
+../../../common/darwin/Classes/VideoOutputManager.swift
