--- media_kit_video/macos/Classes/ResizableTextureProtocol.swift
@@ -1,8 +1,0 @@
-import Cocoa
-import FlutterMacOS
-
-public protocol ResizableTextureProtocol: NSObject, FlutterTexture {
-  func resize(_ size: CGSize)
-  func render(_ size: CGSize)
-  func dispose()
-}
