--- media_kit_video/macos/Classes/SafeResizableTexture.swift
@@ -1,47 +1,0 @@
-import Cocoa
-import FlutterMacOS
-
-// This class avoids data race when called from a thread
-public class SafeResizableTexture:
-  NSObject,
-  FlutterTexture,
-  ResizableTextureProtocol
-{
-  private let lock = NSRecursiveLock()
-  private let child: ResizableTextureProtocol
-
-  init(_ child: ResizableTextureProtocol) {
-    self.child = child
-  }
-
-  public func resize(_ size: CGSize) {
-    return locked {
-      return child.resize(size)
-    }
-  }
-
-  public func render(_ size: CGSize) {
-    return locked {
-      return child.render(size)
-    }
-  }
-
-  public func dispose() {
-    return locked {
-      return child.dispose()
-    }
-  }
-
-  public func copyPixelBuffer() -> Unmanaged<CVPixelBuffer>? {
-    return child.copyPixelBuffer()
-  }
-
-  private func locked<T>(do block: () -> T) -> T {
-    lock.lock()
-    defer {
-      lock.unlock()
-    }
-
-    return block()
-  }
-}
