+++ media_kit_video/macos/Classes/common/MediaKitVideoPlugin.swift
@@ -1,0 +1,1 @@
+../../../common/darwin/Classes/MediaKitVideoPlugin.swift
