+++ media_kit_video/macos/Classes/common/ResizableTextureProtocol.swift
@@ -1,0 +1,1 @@
+../../../common/darwin/Classes/ResizableTextureProtocol.swift
