+++ media_kit_video/macos/Classes/common/SafeResizableTexture.swift
@@ -1,0 +1,1 @@
+../../../common/darwin/Classes/SafeResizableTexture.swift
