+++ media_kit_video/macos/Classes/common/VideoOutput.swift
@@ -1,0 +1,1 @@
+../../../common/darwin/Classes/VideoOutput.swift
