+++ media_kit_video/macos/Classes/common/Worker.swift
@@ -1,0 +1,1 @@
+../../../common/darwin/Classes/Worker.swift
